Move `CacheMap.setExpiration` tests to their own file
